Use render_js_init for InstanceSelectorWidget JS initialisation
This is built-in functionality of the WidgetWithScript utility class used by AdminChooser, and saves us having to include the initialisation JS in the template. Doing it this way means that we can call render_html from the StreamField logic to get just the HTML without the JS (and then trigger the JS at the right point in the StreamField lifecycle).